create packer-plugin-sdk directory and begin moving the relevant folders into it.
